move layer_norm to norm
