+++ lmdeploy/turbomind/builders/__init__.py
@@ -7,11 +7,10 @@ from .attention import AttentionBuilder
 from .decoder_layer import DecoderLayerBuilder, DecoderLayerConfig
 from .deltanet import DeltaNetBuilder
 from .ffn import FfnBuilder, fuse_w1w3
-from .layer_norm import LayerNormBuilder, make_layer_norm_config
 from .mla import MLABuilder
 from .module_list import ModuleListBuilder, ModuleListConfig
 from .moe import MoeBuilder
-from .norm import NormBuilder, make_norm_config
+from .norm import LayerNormBuilder, NormBuilder, make_layer_norm_config, make_norm_config
 from .text_model import TextModelBuilder
 from .vision_model import VisionModelBuilder
